README.md: change **foo** to <i>foo</i>
Since each line in the unordered list of filenames
is a definition, we'd ideally use <dfn>foo</dfn>.
Because <dfn> is sanitized by GitHub's Markdown parser,
we use <i> instead.
This is because <i> is more HTML5 semantic in that
it represents technical names to be read
in a different voice.

Also add this explanation as
a comment in the file itself,
so that future writers
know what to do.